This commit changes SecureMemory to use a single secret and secret factory & adds a new configuration option that can be used to disable mlock and determine whether we want to use the SecureMemory impl or the ProtectedMemory impl. Adding a new option instead of an a new value to secureHeapEngine because we still use mmap to provide all the other protections.
