+++ .github/workflows/ci.yml
@@ -292,7 +292,7 @@ jobs:
     runs-on: ubuntu-latest
     container:
       image: mcr.microsoft.com/dotnet/sdk:3.1
-      options: --ulimit core=-1:-1 --ulimit memlock=-1:-1
+      options: --ulimit core=-1 --ulimit memlock=-1:-1
     steps:
     - name: Checkout the repository
       uses: actions/checkout@5a4ac9002d0be2fb38bd78e4b4dbde5606d7042f
