+++ csharp/SecureMemory/SecureMemory/SecureMemory.csproj
@@ -26,6 +26,8 @@
     </PackageReference>
     <PackageReference Include="System.Diagnostics.Debug" Version="4.3.0" />
     <PackageReference Include="System.Runtime.InteropServices" Version="4.3.0" />
+    <PackageReference Include="System.Configuration.ConfigurationManager" Version="4.4.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="5.0.0" />
     <PackageReference Include="System.Threading" Version="4.3.0" />
   </ItemGroup>
