+++ csharp/SecureMemory/SecureMemory/SecureMemorySecretFactory.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Configuration;
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Runtime.InteropServices;
@@ -6,11 +7,13 @@ using GoDaddy.Asherah.SecureMemory.ProtectedMemoryImpl.Linux;
 using GoDaddy.Asherah.SecureMemory.ProtectedMemoryImpl.MacOS;
 using GoDaddy.Asherah.SecureMemory.ProtectedMemoryImpl.Windows;
 using GoDaddy.Asherah.SecureMemory.SecureMemoryImpl;
+using GoDaddy.Asherah.SecureMemory.SecureMemoryImpl.Linux;
+using GoDaddy.Asherah.SecureMemory.SecureMemoryImpl.MacOS;
 using Microsoft.Extensions.Configuration;
 
-namespace GoDaddy.Asherah.SecureMemory.ProtectedMemoryImpl
+namespace GoDaddy.Asherah.SecureMemory
 {
-    public class ProtectedMemorySecretFactory : ISecretFactory
+    public class SecureMemorySecretFactory : ISecretFactory
     {
         // Detect methods should throw if they know for sure what the OS/platform is, but it isn't supported
         // Detect methods should return null if they don't know for sure what the OS/platform is
@@ -19,7 +22,7 @@ namespace GoDaddy.Asherah.SecureMemory.ProtectedMemoryImpl
         private static object allocatorLock = new object();
         private readonly IConfiguration configuration;
 
-        public ProtectedMemorySecretFactory(IConfiguration configuration)
+        public SecureMemorySecretFactory(IConfiguration configuration)
         {
             Debug.WriteLine("ProtectedMemorySecretFactory ctor");
             lock (allocatorLock)
@@ -49,12 +52,12 @@ namespace GoDaddy.Asherah.SecureMemory.ProtectedMemoryImpl
 
         public Secret CreateSecret(byte[] secretData)
         {
-            return new ProtectedMemorySecret(secretData, allocator, configuration);
+            return new SecureMemorySecret(secretData, allocator, configuration);
         }
 
         public Secret CreateSecret(char[] secretData)
         {
-            return ProtectedMemorySecret.FromCharArray(secretData, allocator, configuration);
+            return SecureMemorySecret.FromCharArray(secretData, allocator, configuration);
         }
 
         public void Dispose()
@@ -81,6 +84,103 @@ namespace GoDaddy.Asherah.SecureMemory.ProtectedMemoryImpl
                     Debug.WriteLine($"ProtectedMemorySecretFactory: New refCount is {refCount}");
                 }
             }
+        }
+
+        internal static ISecureMemoryAllocator ConfigureForMacOS64(IConfiguration configuration)
+        {
+            if (configuration != null)
+            {
+                string secureHeapEngine = configuration["secureHeapEngine"];
+                string mLock = configuration["mlock"];
+                if (!string.IsNullOrWhiteSpace(secureHeapEngine))
+                {
+                    if (string.Compare(secureHeapEngine, "openssl11", StringComparison.InvariantCultureIgnoreCase) == 0)
+                    {
+                        throw new PlatformNotSupportedException(
+                            "OpenSSL 1.1 selected for secureHeapEngine but is not yet supported for MacOS");
+                    }
+
+                    if (string.Compare(secureHeapEngine, "mmap", StringComparison.InvariantCultureIgnoreCase) == 0)
+                    {
+                        if (!string.IsNullOrWhiteSpace(mLock))
+                        {
+                            if (string.Compare(mLock, "disabled", StringComparison.InvariantCultureIgnoreCase) == 0)
+                            {
+                                return new MacOSSecureMemoryAllocatorLP64();
+                            }
+
+                            throw new ConfigurationErrorsException("Unknown mlock configuration: " + mLock);
+                        }
+
+                        return new MacOSProtectedMemoryAllocatorLP64();
+                    }
+
+                    throw new PlatformNotSupportedException("Unknown secureHeapEngine: " + secureHeapEngine);
+                }
+
+                if (!string.IsNullOrWhiteSpace(mLock))
+                {
+                    if (string.Compare(mLock, "disabled", StringComparison.InvariantCultureIgnoreCase) == 0)
+                    {
+                        return new MacOSSecureMemoryAllocatorLP64();
+                    }
+
+                    throw new ConfigurationErrorsException("Unknown mlock configuration: " + mLock);
+                }
+            }
+
+            return new MacOSProtectedMemoryAllocatorLP64();
+        }
+
+        internal static ISecureMemoryAllocator ConfigureForLinux64(IConfiguration configuration)
+        {
+            if (configuration != null)
+            {
+                string secureHeapEngine = configuration["secureHeapEngine"];
+                string mLock = configuration["mlock"];
+                if (!string.IsNullOrWhiteSpace(secureHeapEngine))
+                {
+                    if (string.Compare(secureHeapEngine, "openssl11", StringComparison.InvariantCultureIgnoreCase) == 0)
+                    {
+                        if (LinuxOpenSSL11ProtectedMemoryAllocatorLP64.IsAvailable())
+                        {
+                            return new LinuxOpenSSL11ProtectedMemoryAllocatorLP64(configuration);
+                        }
+
+                        throw new PlatformNotSupportedException(
+                            "OpenSSL 1.1 selected for secureHeapEngine but library not found");
+                    }
+
+                    if (string.Compare(secureHeapEngine, "mmap", StringComparison.InvariantCultureIgnoreCase) == 0)
+                    {
+                        if (!string.IsNullOrWhiteSpace(mLock))
+                        {
+                            if (string.Compare(mLock, "disabled", StringComparison.InvariantCultureIgnoreCase) == 0)
+                            {
+                                return new LinuxSecureMemoryAllocatorLP64();
+                            }
+
+                            throw new ConfigurationErrorsException("Unknown mlock configuration: " + mLock);
+                        }
+
+                        return new LinuxProtectedMemoryAllocatorLP64();
+                    }
+
+                    throw new PlatformNotSupportedException("Unknown secureHeapEngine: " + secureHeapEngine);
+                }
+
+                if (!string.IsNullOrWhiteSpace(mLock))
+                {
+                    if (string.Compare(mLock, "disabled", StringComparison.InvariantCultureIgnoreCase) == 0)
+                    {
+                        return new LinuxSecureMemoryAllocatorLP64();
+                    }
+
+                    throw new ConfigurationErrorsException("Unknown mlock configuration: " + mLock);
+                }
+            }
+
+            return new LinuxProtectedMemoryAllocatorLP64();
         }
 
         [ExcludeFromCodeCoverage]
@@ -179,61 +279,6 @@ namespace GoDaddy.Asherah.SecureMemory.ProtectedMemoryImpl
 
             // We return null if we don't know what the OS is, so other methods can be tried
             return null;
-        }
-
-        private static ISecureMemoryAllocator ConfigureForLinux64(IConfiguration configuration)
-        {
-            if (configuration != null)
-            {
-                var secureHeapEngine = configuration["secureHeapEngine"];
-                if (!string.IsNullOrWhiteSpace(secureHeapEngine))
-                {
-                    if (string.Compare(secureHeapEngine, "openssl11", StringComparison.InvariantCultureIgnoreCase) == 0)
-                    {
-                        if (LinuxOpenSSL11ProtectedMemoryAllocatorLP64.IsAvailable())
-                        {
-                            return new LinuxOpenSSL11ProtectedMemoryAllocatorLP64(configuration);
-                        }
-
-                        throw new PlatformNotSupportedException(
-                            "OpenSSL 1.1 selected for secureHeapEngine but library not found");
-                    }
-
-                    if (string.Compare(secureHeapEngine, "mmap", StringComparison.InvariantCultureIgnoreCase) == 0)
-                    {
-                        return new LinuxProtectedMemoryAllocatorLP64();
-                    }
-
-                    throw new PlatformNotSupportedException("Unknown secureHeapEngine: " + secureHeapEngine);
-                }
-            }
-
-            return new LinuxProtectedMemoryAllocatorLP64();
-        }
-
-        private static ISecureMemoryAllocator ConfigureForMacOS64(IConfiguration configuration)
-        {
-            if (configuration != null)
-            {
-                var secureHeapEngine = configuration["secureHeapEngine"];
-                if (!string.IsNullOrWhiteSpace(secureHeapEngine))
-                {
-                    if (string.Compare(secureHeapEngine, "openssl11", StringComparison.InvariantCultureIgnoreCase) == 0)
-                    {
-                        throw new PlatformNotSupportedException(
-                            "OpenSSL 1.1 selected for secureHeapEngine but is not yet supported for MacOS");
-                    }
-
-                    if (string.Compare(secureHeapEngine, "mmap", StringComparison.InvariantCultureIgnoreCase) == 0)
-                    {
-                        return new MacOSProtectedMemoryAllocatorLP64();
-                    }
-
-                    throw new PlatformNotSupportedException("Unknown secureHeapEngine: " + secureHeapEngine);
-                }
-            }
-
-            return new MacOSProtectedMemoryAllocatorLP64();
         }
 
         [ExcludeFromCodeCoverage]
